Move acquisition test context config to separate file.
